Add CHECKDATASIG to script standard flags (#1750)
* Remove redundant items from STANDARD_SCRIPT_VERIFY_FLAGS

SCRIPT_VERIFY_LOW_S and SCRIPT_VERIFY_NULLFAIL are already included
as part of MANDATORY_SCRIPT_VERIFY_FLAGS.

(see https://reviews.bitcoinabc.org/D3059)

* Clarify comment for SCRIPT_ENABLE_CHECKDATASIG

The important function of the flag, now that it is activated,
is to flag when it should be counted towards the sigops limit.

* Make tests independent of whether CHECKDATASIG is included in mandatory or standard flags

Currently, in addition to testing the functioning of CHECKDATASIG,
sigopcount_tests.cpp and checkdatasig_tests.cpp implicitly test that
CHECKDATASIG is not included in standard or mandatory flags.

Change this to make the CHECKDATASIG tests independent of whether CHECKDATASIG
is included in standard or mandatory flags. To maintain test coverage, also
add a test to explicitly test whether CHECKDATASIG is included in standard or
mandatory flags.

(port https://reviews.bitcoinabc.org/D3063)

* Add CHECKDATASIG to standard flags.

Also remove STANDARD_CHECKDATASIG_VERIFY_FLAGS.

(port https://reviews.bitcoinabc.org/D3069)

* Interpret scripts with CHECKDATASIG opcode always valid.

Because CHECKDATASIG uses a previouly invalid opcode number, it does
not appear in any executed branches of scripts prior its introduction in the
Nov 15, 2018 Upgrade. This means that the interpreter can treat it as if
it has always been valid.

Before this change, the SCRIPT_ENABLE_CHECKDATASIG flag had both a permissive
(enabling the opcode) and a restrictive (enforcing siops counting) meaning.
This change simplifies the meaning of the flag to be restrictive only.

(port of https://reviews.bitcoinabc.org/D3100)