Commit the new txns to the txindex database in ConnectBlock()
WriteBestBlock in BlockConnected()

This is to ensure that we've already synced the txindex and we
don't end up writing a better block than we updated the txindex
to.  Generally it wouldn't be an issue but if someone restarts
bitcoind before the txindex is fully up to date then they could
end up with an invalid best block and end up skipping many
blocks.

Also, remove SetBestChain as it wasn't being used and is not
necessary in BU.