Use a set to ensure we don't double count peers (#1821)
* Use a set to ensure we don't double count peers

We can receive multiple VERSION or SENDCMPCT messages which can
cause us to double count those peer types. By using a set
we can easily filter out the duplicates.

* Release the node reference

When we're done setting up the request of a transaction we can
finally release the ref and delete the node from the items list.

* Do not allow the sets for counting thin peers to grow unbounded

If somehow the sets get larger than the number of nodes connected
then assert (if in debug) and do nothing and return if running
a non debug build.