MINIMALDATA consensus activation
Summary:
This converts MINIMALDATA from standard to consensus for the Nov 2019
upgrade. The behaviour of MINIMALDATA and the reasons for activation
are described here: https://github.com/bitcoincashorg/bitcoincash.org/pull/376

Test is based on D3736 but doesn't test upgrade deactivation since this is
a strict soft-fork.

Note that in validation.cpp CheckInputs, mandatoryFlags never includes
MINIMALDATA so minimaldata violations always appear as "non-mandatory-...",
even after the upgrade, not as "upgrade-conditional-...".

Differential Revision: D3763