Make it so that uacomment[] are correctly reported in subver string (#1830)
* Make it so that uacomment[] are correctly reported in subver string.

This is a fix for issue #1601. Namely "uaccoments" parameters were
overwritten by EB/AD settings. These change are always appending
"uaccoments" to subver string in the init step and also at run time
when EB and/or AD are changed.

If the resulting subver is longer than 256 chars (per BIP 014), in the
init step we are going to emit a warning and exit. If this situation triggers
at run time we just trim the subver until it met the 256 chars limit.

* Test subver string for "uaccoment" and length

* Add a 32bit identifier to the uacomment string

This will help us know whether we still need to support 32bit
releases on mainnet.

* Make strSubVersion local to init.cpp

* Warn about excessive subver length in FormatSubVersion()

Moving the check of lenght and trimming in FormatSubVersion(),
make it so that the check for lenght in init.cpp can never fail and
InitError() never trigger.

Wiithout this change useisr would never have been notified about setting
a uacomment param that produce a too long subver string.

* Remove truncation version from FormatSubVersion and it back to init.cpp

While doing that use InitWarning rather than a LOGA, since we are doing
the check during bootstrap and that uacomment can't be modified at runtime.

NOTE: This solution isn't perfect because the user could have entered in
uacomments which exactly hits the max length without the need for truncation.
To be able to determine if truncation was actually performed we'd have to
modify FormatSubVersion to be able to return this information in addition
to the strSubVersion.

* Remove uacomment sanitization from FormatSubVersion()

This sanitisation is performed during the bootstrap phase (init.cpp)
and since uacomments can't be changed at runtime, the removed check
was redundant.

* Add arch identifier just after EB/AD, rather than at the end of  subver.

* Fix an off-by-one error while trimming subver string

* Check that subver is correctly trimmed at 256 chars length

* Make the unit test for FormatSubVersion to handle different architecture properly

* Reset back EB/AD to default values

* Tweak to turn on/off the display of node architecture on subver string