remove SCRIPT_ENABLE_SCHNORR flag and clean up tests
remove SCRIPT_ENABLE_SCHNORR flag along with a bunch of duplicate
tests now that the flag is impotent.

In addition to removing those tests, there were numerous other script_tests
cases that are now redundant as well, also removed in this diff.

This a port of https://reviews.bitcoinabc.org/D3332