Refactor to return all txn rate stats atomically as a group
This change serves two purposes:
1. Group the return of all transaction rate statistics in a single
method in order to reduce the number of times the cs_txPerSec lock is
taken.  These statistics are commonly requested together and the old
method of individually returning each stat would cause frequent taking
and releasing of the lock in quick succession.  By grouping the return
under a single lock, this also ensures that the statistics are all
returned atomically.  As part of this, the individual stat return
methods have been removed to discourage misuse. This is necessary as the
stats will be retrieved at regular intervals now for display in the
transaction rate graph (to be added in subsequent commit).

2. Allow updating transaction rate stats without adding a new
transaction.  This is to allow continual decay of the smoothed
transaction rate and re-computation of the instantaneous rate during
periods of time longer than the instantaneous resolution when no
transactions are received.  Previously if no transactions were received
the smoothed transaction rate would incorrectly show as plateaued at the
rate computed when the last transaction was received.

This issue is much more obvious when viewed in a trend chart (subsequent
commit in this feature adds that trend chart to the Qt GUI).