Expose some local variables in the update txn rate method
The mempool computes an instantaneous transaction rate at 1-second
resolution in order to update the peak transaction rate for the entire
runtime of the node.  This instantaneous rate was being thrown away
after updating the peak.  This commit exposes the most recent
instantaneous rate to be used for charting.
Additionally exposes the smoothing duration and instantaneous transaction
resolution constants for external use.