CPU Miner Enhancements -- Follow-up
As discussed in PR #1904, it was a bad idea to pass down a
shared_ptr<CReserveScript> to mkblocktemplate() from the
getminingcandidate RPC handler. Doing it that way was not terribly useful
behavior and was just needlessly noisy to read.

(The only reason we might want to do it that way would have been if we
wanted to call 'KeepKey()' on the CReserveScript reference returned from
the wallet -- but we never did that -- so it was pointless to use a
shared_ptr).

So, in the interests of more readable/simpler code, this commit passes
down a simple const CScript & (optional argument) to mkblocktemplate.

Also in this commit: Updated the comments and return the correct RPC
error messages for the two error cases involving the wallet:

- wallet not compiled in -> "No coinbase script available (mining requires a wallet)"
- wallet has no keypool keys -> "Error: Keypool ran out, please call keypoolrefill first"

Previously the code actually reversed those two error messages: It would
return the error for no more keys if no wallet support compiled-in, and
return the "no wallet support" error message for when the keypool ran out
of keys!