[SECP256K1] Create a different library when building with JNI
Summary:
The current behavior is to build a `secp256k1` library that has the JNI
binding feature if enabled.
This diff keeps the main `secp256k1` library intact, and creates another
`secp256k1_jni` library with JNI support in it. The java test is updated
to reflect this.

Depends on D4053.

Test Plan:
  cd secp256k1
  ./autogen.sh
  mkdir build && cd build
  ../configure --enable-jni --enable-experimental --enable-module-ecdh
  make
Check there is a `libsecp256k1_jni.so` file under `.libs/`
  make check-java

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4058