Don't run CheckBlock() more than once during IBD (#2057)
* Don't run CheckBlock() more than once during IBD

CheckBlock() is generally limited from running more than once by
using the fChecked flag which is stored internally to the block however
during IBD we store the blocks to disk 1000 blocks in advance which
ends up clearing the flag so when the block is read again from disk
we end up doing CheckBlock() again. This is an exepensive operation
and takes up 10% of the total time for IBD to run. So by temporarily
storing the hash of any block we have run CheckBlock() on during IBD
we can then avoid re-running CheckBlock() after that block is read
again from disk.

* Add critical section for setBlocksAlreadyChecked

* For efficiency use an unordered_set

* Do not hold cs_BlocksAlreadyChecked when doing CheckBlock()