[Giga-Net:] Add a priority queue for urgent message processing (#1721)
* Add a priority message receive queue and use it

Swap the first message in queue rather than copy

  In the past, due to the need to keep the lock a minimal amount of time
  a full copy of the message was being done. The purpose of this was
  to allow for better multi-threading, however, this is not efficient
  especially with potentially large messages. Instead, due a swap
  which requires no memory re-allocations. The trick here is we "must" do
  the swap only after we are certain we have a complete message or we
  will crash.  We can do this simply by checking for the complete message
  a little earlier thus allowing the swap to safely continue.

Do not invoke priority queuing unless the chain is nearly syncd

  Priority messaging is only needed when we are propagating thin type blocks
  or headers which are propagated as new block announcements.

Do not swap the first message as it could have been partially sent

  Send the first two messages in queue because we had to place the
  priority message second in case the first message was already partially
  sent.

Re-enable asserts ih SocketSendData() but change them to DbgAsserts()

Leave the priority message at front of the queue when nSendOffset==0

  In the case of nSendOffset == 0, we know that the message
  that was at the front of the queue before we added the priority message
  is still entact and has not been partially sent. This will be in the
  vast majority of cases true. Therefore we can just leave the priority
  message at the front and it will be the first message sent.

  In the rare event where the first message had been parially sent we
  can leave it at the front and move the first priority message behind it.

Add Compact Block messages to the priority queuing

  We can add most of them except for the getdata() which would
  require us to deserialize every getdata message. It would be better
  if we upgraded compact blocks using XVERSION so that we could add
  a separate p2p message "get_cmpct" and then add that to the priority
  queing.

Do not add a priority noderef back to the queue if fDisconnect is true

  Prevent infinite looping if there is a socket issue or network congestion

Add note why BLOCK as a priority message type is not included.

Only allow up to 5 simultaneous priority inbound messages

  We don't want to allow some peer to dominate our networking by
  sending us a stream of priority messages. There should in general be
  only one priority message in the queue with the exception of HEADERS coming
  from block announcements, which could arrive relatively close in time.

Make sure to erase any noderefs that may be left in the priority send queue

  When the socket is closed there may still be priority messages waiting
  to be sent. Since these messages use a noderef then our noderef count
  will never get to zero and the disconnect will not finalize unless
  we delete them just after the socket is closed.

* Add tests for the priority send queue

* Replace datastream insert using char* with std::string().

  Using the insert fuction using char* is prone to failure at the
  boundary where we try to insert the last char of a string but can't
  because the vector insert will only insert up to that value but
  not the value itself. So by replacing the function to usea string iterators
  we can pass the end() iterator which allow us to insert all the values
  up to the end of the string.

  Since this function is only used in two places, in test code, it
  has little impact.

* Simplify receive message procesing.

Don't add message chunks to the back of the vRecvMsg queue but rather
keep a temporary message which we can add bytes to and then when complete
we can push it onto either the prioirty queue or the normal queue by
using std::swap and std::move. This makes the code simpler and easier
to understand.

* Remove check for incomplete message

* Do not sleep if a priority message is in queue

* Prevent possible infinite loop

If somehow fPriorityRecvMsg was true but the queue was empty then
we could end up in a loop when we try to access the next message
from the priorityq.  Set the priorty flag to false in such an instance
and then continue.

* Add New Priority Send Queue

* Fix exploit_tests

using different iterators for comparison seems to cause
some compilers trouble.

* Improve looping performance where we grab the inbound priority msg's

If the try lock fails there is no need to continue trying to get
the lock on the priority message queue...clearly some other thread
is processing the priority messages and we can just move on and
continue processing the lower priority messaegs in our thread.

* FIx bug in receive queue