[travis] Remove constraints on build time for dependencies and… (#2096)
Since Travis raised the default job timeout threshold to 180 minute
we don't need to exit early in the to save dependencies and compiler
cache. Unfortunately this work only for
BitcoinUnlimited/BitcoinUnlimited repo, so we need to special case it to
avoid to get timeout on forks of the BU codebase.