[Minor enhancement]  Startup txindex after ActivateBestChain (#2094)
* Startup txindex near the end of the node initialization

In a case where an IBD is halted then restarted we can end up having
a great deal of txindex compaction to do which can slow down node
startup significantly.  If we start txindex near the end, we will
make the node available much faster which txindex runs in the
background.

* Check for valid g_txindex pointer rather than fTxIndex flag

We're going to use the pointer so we should check for the pointer
validitity rather than relying on whether the fTxIndex flag is set
correctly. Also we can check for pointer validity by using
IsTxIndexReady().

* Do not start the txindexing process until the node has fully synced

* Remove log spam when checking synced status

* Clarify the code and add comments when we startup the txindex database

We don't want to rely of the fReindex flag to determine whether to wipe
and rebuild the txindex because, in the case of a very small reindex, it
could already be complete by the time we get to the end of the node
startup where we initialize the txindex, and therefore the txindex would
not get rebuilt. So instead we determine whether the config parameters
had asked for a -reindex and if so then also rebuild the txindex.