Serialize AcceptToMemoryPool()
This function is only used in rpc call "sendrawtransaction" or in the
wallet when sending a transaction.  In order to check properly for
the double spends we need to make sure we add a lock at the beginning
of the function call and then pause and flush the tx commit queue before
attempting to check and accept this transaction.