Make thinblock and grapheneblock checks more permissive
Take out the checking for inv.type when we first start
processing thinblocks or grapheneblocks. These are not really
necessary, and furthermore, this will allow us to remove the
msg_thinblock, msg_xthinblock and msg_grapheneblock enums after
the next hard fork is complete. This will be helpful, especially
in the case of graphene, which is reaching maturity and will
be ported to BCHN in the near future.