change XVersionMap from unorderded_map to map for better performance
In terms of map vs unordered_map there is no performance difference at
the datastructure because the data set is so small.

map has two advantages over unordered_map. Its keys are always sorted so
we do not leak any information about system entropy thrugh the hasher.
The use of secure random numbers in the map hasher is expensive and
these are not needed by std::map.