DSProof: Various nits and fixups
Summary of changes:

- Use proper bracketed includes as is done in this codebase (e.g.
  #include <thing.h> rather than #include "thing.h"
- Header guards must be prefixed with BITCOIN_ (as is standard in this
codebase)
- Removed unused #includes and moved around misplaced #includes and
added missing #includes
- Alphabetized some includes
- DoubleSpendProof.prevTxId() made to return by reference, as a
performance optimization. All call sites were checked to ensure this is
safe.
- firstSpender() and doubleSpender() made to return a reference to save
a copy. These methods aren't called anywhere but if they are to be used
in the future, it's best to avoid extra copying of data.
- Made a few getters inline
- Got rid of a #define in favor of static constexpr
- We don't use the std::mutex, std::recursive_mutex types. Instead we
use Mutex and RecursiveMutex because they are compatible with thread
safety annotations.  Renamed std::recursive_mutex -> RecursiveMutex
- Got rid of a bad #ifdef NDEBUG which would fail if compiling with
NDEBUG but with assert (not a problem for this codebase really).
- Got rid of the HashShortener type as this hasher is unsafe due to the
possibility of hash table collision attacks.  Instead we use a salted
hasher based off the SaltedTxidHasher used in txmempool.h
- Got rid of the bare DSProofStorage pointer in favor of a unique_ptr