Remove orphans (if any) when adding a new dsproof
There is a possiblity due to multi-threading that we could add
a new orphan during the time we're validating a dsproof which
could result in a dsproof getting added which is fully validated
but is marked as an orphan. This is caused by locks which can
not be held throughout the process of detecting and triggering
on a doublespend and subsequently creating a new dsproof.