Fix bug in connection logic and tweak ref
When we try to connect automatically to a default number of
xthin nodes and then we update the tweak value, the tweak value
isn't being used. This is because the value of minXthinNodes is
getting it's value reassigned on each loop using GetArg() which
then also reassigns the tweak ref value.  To fix this we
assign the value using GetArg() just once at the beginning of
the connection loop.

This brings up another issue of retiring old config options and just
using tweaks but for now that is beyond the scope of this commit.