Update hashBlock in GetTransaction() if needed
Previously we didn't update hashBlock in case we actually found the tx
included in the block stored on disk corresponding to the hash parameter
passed to getrawtransaction rpc.

That meant that in case the node was running with txindex=0, getrawtransaction
output would have considered the confirmed transaction like it belongs to the
mempool rather than to a block stored on disk, i.e. confirmation and time fields
would have been set to 0 and txTime respectively, whereas blocktime field would
has been dropped.