Disconnect peers if they give us a bad header at a checkpoint
These peers will typcially be ones that were forked off the network
and if we don't disconnect them then we can get stuck during IBD
if/when we download the initial headers. If for example, the first
three peers to connect to ours are one the wrong fork then we'll
only download headers up to the last fork.