Fix GuessVerificationProgress when checkpoints are enabled.
If checkpoints are enabled then we don't need to account for signature
verification during IBD.

Pass the value of !fCheckpointsEnabled to enable/disable sig checks.

However in the wallet when we do a rescan we don't need to account
for sigchecks so make sure that all calls to GuessVerificationProgress
are passed a "false". There was still one instance where we were not
doing that.